+++ CHANGELOG.md
@@ -13,6 +13,7 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 - New `db:build:ext` command to generate a draft (json format) of each components in one extension.
 - New `db:list` command to see what are extensions supported by the database.
 - New `ExtensionFactory::getExtensions()` method to retrieve all extensions informations (status/versions)
+- New `db:show` command to see details of extensions supported by the database.
 
 ### Changed